refactored test file structure to match lib, removed plurals, and fix… (#89)
* refactored test file structure to match lib, removed plurals, and fixed imports

* removed duplicate import

* renamed widget tests to include _widget_test.dart suffix

* restored "as constant" that was lost in the rebase/conflict resolution process

* restored go_to_page.dart (old)

* goto_page_widget_test.dart is not yet compatible with goto_page.dart (new)

* removed unused and flutter format

* test imports

* fixed confusion between new and old goto pages

* fixed minor warnings and ran flutter format

* highlighting widget test is for goto_page_old.dart
