+++ lib/model/segment.dart
