+++ lib/view/home_page.dart
@@ -1,0 +1,18 @@
+import 'package:concordi_around/widget/drawer.dart';
+import 'package:concordi_around/widget/map.dart';
+import 'package:flutter/material.dart';
+
+class HomePage extends StatefulWidget {
+  @override
+  State<StatefulWidget> createState() {
+    return _HomePageState();
+  }
+}
+
+class _HomePageState extends State<HomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        body: Map(), drawer: SidebarDrawer(), resizeToAvoidBottomInset: false);
+  }
+}
