+++ lib/widget/svg_floor_plan/enter_building_action_button.dart
