+++ lib/widget/svg_floor_plan/floor_selector_enter_building_column.dart
@@ -1,0 +1,73 @@
+import 'package:concordi_around/provider/map_notifier.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import 'enter_building_action_button.dart';
+
+class FloorSelectorEnterBuilding extends StatefulWidget {
+  final Function(int) selectedFloor;
+  final VoidCallback enterBuildingPressed;
+
+  FloorSelectorEnterBuilding({this.selectedFloor, this.enterBuildingPressed});
+
+  @override
+  State<StatefulWidget> createState() {
+    return _FloorSelectorEnterBuildingState();
+  }
+}
+
+class _FloorSelectorEnterBuildingState
+    extends State<FloorSelectorEnterBuilding> {
+  List<bool> selectedFloor = [true, false]; // 9th floor selected
+
+  Widget build(BuildContext context) {
+    MapNotifier mapNotifier = Provider.of<MapNotifier>(context);
+
+    return Padding(
+      padding: EdgeInsets.fromLTRB(16, 0, 0, 16),
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.end,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: <Widget>[
+          Visibility(
+            maintainSize: true,
+            maintainAnimation: true,
+            maintainState: true,
+            visible: mapNotifier.showFloorPlan,
+            child: Container(
+              child: RotatedBox(
+                quarterTurns: 1,
+                child: Container(
+                  decoration: BoxDecoration(
+                      color: Colors.white,
+                      borderRadius: BorderRadius.circular(20)),
+                  child: ToggleButtons(
+                    selectedColor: Colors.white,
+                    color: Colors.black,
+                    fillColor: Colors.grey,
+                    borderRadius: BorderRadius.circular(20),
+                    borderColor: Colors.grey,
+                    children: <Widget>[
+                      RotatedBox(quarterTurns: 3, child: Text("9")),
+                      RotatedBox(quarterTurns: 3, child: Text("8")),
+                    ],
+                    isSelected: selectedFloor,
+                    onPressed: (int index) {
+                      widget.selectedFloor(index == 0 ? 9 : 8);
+                      setState(() {
+                        selectedFloor = [false, false];
+                        selectedFloor[index] = true;
+                      });
+                    },
+                  ),
+                ),
+              ),
+            ),
+          ),
+          SizedBox(height: 16, width: 16),
+          EnterBuildingButton(isPressed: () => {widget.enterBuildingPressed()}),
+        ],
+      ),
+    );
+  }
+}
