--- lib/widgets/custom/custom_fade_indexed_stack.dart
@@ -1,63 +1,0 @@
-import 'package:flutter/animation.dart';
-import 'package:flutter/widgets.dart';
-
-class FadeIndexedStack extends StatefulWidget {
-  final int index;
-  final List<Widget> children;
-  final Duration duration;
-
-  const FadeIndexedStack({
-    Key key,
-    this.index,
-    this.children,
-    this.duration = const Duration(
-      milliseconds: 750,
-    ),
-  }) : super(key: key);
-
-  @override
-  _FadeIndexedStackState createState() => _FadeIndexedStackState();
-}
-
-class _FadeIndexedStackState extends State<FadeIndexedStack>
-    with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-
-  @override
-  void didUpdateWidget(FadeIndexedStack oldWidget) {
-    if (widget.index != oldWidget.index) {
-      _controller.forward(from: 0.0);
-    }
-    super.didUpdateWidget(oldWidget);
-  }
-
-  @override
-  void initState() {
-    _controller = AnimationController(vsync: this, duration: widget.duration);
-    _controller.forward();
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedBuilder(
-      animation: _controller,
-      builder: (context, child) {
-        return Opacity(
-          opacity: _controller.value,
-          child: child,
-        );
-      },
-      child: IndexedStack(
-        index: widget.index,
-        children: widget.children,
-      ),
-    );
-  }
-}
